Fix output encoding.
Outputs from the program should be binary encoded.  Without this, python will
do the best it can and try to coerce the output to binary and implicitly encode
it with whatever the locale says.  For many people this is ascii, and for many
people this breaks.

Being explicit about the output encoding gets the script moving again.